requirements.txt
alembic==1.6.5
click==8.0.1
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.10
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.19.5
pandas==1.2.5
Pillow==8.2.0
psycopg2==2.8.4
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.6
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.19
threadpoolctl==2.1.0
Werkzeug==2.0.1
